refactor: Straight Litzwire 및 가공위치 정보 ProjectStraightEntity 로 통합 처리
